move panel-resize service to a new folder
