+++ core/gui/src/app/workspace/component/result-panel/result-panel.component.ts
@@ -5,7 +5,8 @@ import { WorkflowActionService } from "../../service/workflow-graph/model/workfl
 import { ExecutionState, ExecutionStateInfo } from "../../types/execute-workflow.interface";
 import { ResultTableFrameComponent } from "./result-table-frame/result-table-frame.component";
 import { ConsoleFrameComponent } from "./console-frame/console-frame.component";
-import { WorkflowResultService, PanelResizeService } from "../../service/workflow-result/workflow-result.service";
+import { WorkflowResultService } from "../../service/workflow-result/workflow-result.service";
+import {PanelResizeService} from "../../service/workflow-result/panel-resize/panel-resize.service";
 import { VisualizationFrameComponent } from "./visualization-frame/visualization-frame.component";
 import { filter } from "rxjs/operators";
 import { UntilDestroy, untilDestroyed } from "@ngneat/until-destroy";
