+++ core/gui/src/app/workspace/component/result-panel/result-table-frame/result-table-frame.component.ts
@@ -6,8 +6,8 @@ import { WorkflowActionService } from "../../../service/workflow-graph/model/wor
 import {
   DEFAULT_PAGE_SIZE,
   WorkflowResultService,
-  PanelResizeService,
 } from "../../../service/workflow-result/workflow-result.service";
+import {PanelResizeService} from "../../../service/workflow-result/panel-resize/panel-resize.service";
 import { isWebPaginationUpdate } from "../../../types/execute-workflow.interface";
 import { IndexableObject, TableColumn } from "../../../types/result-table.interface";
 import { RowModalComponent } from "../result-panel-modal.component";
