+++ core/gui/src/app/workspace/service/workflow-result/panel-resize/panel-resize.service.ts
@@ -1,0 +1,14 @@
+import { Injectable } from "@angular/core";
+import { BehaviorSubject } from "rxjs";
+
+@Injectable({
+    providedIn: "root",
+  })
+  export class PanelResizeService {
+    private panelSizeSource = new BehaviorSubject<{ width: number; height: number }>({ width: 800, height: 300 });
+    currentSize = this.panelSizeSource.asObservable();
+  
+    changePanelSize(width: number, height: number) {
+      this.panelSizeSource.next({ width, height });
+    }
+  }
